Adding album API support
