+++ cmd/getapplemusicmedia/getapplemusicmedia.go
@@ -1,12 +1,9 @@
 package getapplemusicmedia
 
 import (
-	"context"
 	"encoding/json"
-	"fmt"
 	"log"
 	"net/http"
-	"os"
 
 	"cloud.google.com/go/firestore"
 	"github.com/pixelogicdev/gruveebackend/pkg/firebase"
@@ -70,206 +67,41 @@ func GetAppleMusicMedia(writer http.ResponseWriter, request *http.Request) {
 		return
 	}
 
+	var (
+		firestoreMediaData    interface{}
+		firestoreMediaDataErr error
+	)
+
 	// Time to make our request to Apple Music API
 	switch appleMusicMediaReq.MediaType {
 	case "track":
-		// Call track API
-		firestoreMediaTrackData, firestoreMediaDataErr := getAppleMusicTrack(appleMusicMediaReq.MediaID, appleMusicMediaReq.Storefront, *appleDevToken)
+		firestoreMediaData, firestoreMediaDataErr = getAppleMusicTrack(appleMusicMediaReq.MediaID, appleMusicMediaReq.Storefront, *appleDevToken)
 		if firestoreMediaDataErr != nil {
 			http.Error(writer, firestoreMediaDataErr.Error(), http.StatusInternalServerError)
-			log.Printf("GetSpotifyMedia [GetSpotifyTrack Switch]: %v", firestoreMediaDataErr)
+			log.Printf("GetAppleMusicMedia [GetAppleMusicTrack Switch]: %v", firestoreMediaDataErr)
 			return
 		}
-
-		writer.WriteHeader(http.StatusOK)
-		writer.Header().Set("Content-Type", "application/json")
-		json.NewEncoder(writer).Encode(firestoreMediaTrackData)
-		return
 	case "playlist":
-		// Call playlist API
-		firestoreMediaPlaylistData, firestoreMediaDataErr := getAppleMusicPlaylist(appleMusicMediaReq.MediaID, appleMusicMediaReq.Storefront, *appleDevToken)
+		firestoreMediaData, firestoreMediaDataErr = getAppleMusicPlaylist(appleMusicMediaReq.MediaID, appleMusicMediaReq.Storefront, *appleDevToken)
 		if firestoreMediaDataErr != nil {
 			http.Error(writer, firestoreMediaDataErr.Error(), http.StatusInternalServerError)
-			log.Printf("GetSpotifyMedia [GetSpotifyPlaylist Switch]: %v", firestoreMediaDataErr)
+			log.Printf("GetAppleMusicMedia [GetAppleMusicPlaylist Switch]: %v", firestoreMediaDataErr)
 			return
 		}
-
-		writer.WriteHeader(http.StatusOK)
-		writer.Header().Set("Content-Type", "application/json")
-		json.NewEncoder(writer).Encode(firestoreMediaPlaylistData)
+	case "album":
+		firestoreMediaData, firestoreMediaDataErr = getAppleMusicAlbum(appleMusicMediaReq.MediaID, appleMusicMediaReq.Storefront, *appleDevToken)
+		if firestoreMediaDataErr != nil {
+			http.Error(writer, firestoreMediaDataErr.Error(), http.StatusInternalServerError)
+			log.Printf("GetAppleMusicMedia [GetAppleMusicAlbum Switch]: %v", firestoreMediaDataErr)
+			return
+		}
+	default:
+		http.Error(writer, appleMusicMediaReq.MediaType+" media type does not exist", http.StatusInternalServerError)
+		log.Printf("GetAppleMusicMedia [MediaTypeSwitch]: %v media type does not exist", appleMusicMediaReq.MediaType)
 		return
-		// case "album":
-		// 	// Call album API
-		// 	firestoreMediaAlbumData, firestoreMediaDataErr := getSpotifyAlbum(spotifyMediaReq.MediaID, creds.AccessToken)
-		// 	if firestoreMediaDataErr != nil {
-		// 		http.Error(writer, firestoreMediaDataErr.Error(), http.StatusInternalServerError)
-		// 		log.Printf("GetSpotifyMedia [GetSpotifyPlaylist Switch]: %v", firestoreMediaDataErr)
-		// 		return
-		// 	}
-
-		// 	writer.WriteHeader(http.StatusOK)
-		// 	writer.Header().Set("Content-Type", "application/json")
-		// 	json.NewEncoder(writer).Encode(firestoreMediaAlbumData)
-		// 	return
-		// default:
-		// 	http.Error(writer, spotifyMediaReq.MediaType+" media type does not exist", http.StatusInternalServerError)
-		// 	log.Printf("GetSpotifyMedia [MediaTypeSwitch]: %v media type does not exist", spotifyMediaReq.MediaType)
-		// 	return
 	}
 
-}
-
-// Helpers
-
-// initWithEnv takes our yaml env variables and maps them properly.
-// Unfortunately, we had to do this is main because in init we weren't able to access env variables
-func initWithEnv() error {
-	// Get paths
-	var currentProject string
-
-	if os.Getenv("ENVIRONMENT") == "DEV" {
-		currentProject = os.Getenv("FIREBASE_PROJECTID_DEV")
-	} else if os.Getenv("ENVIRONMENT") == "PROD" {
-		currentProject = os.Getenv("FIREBASE_PROJECTID_PROD")
-	}
-
-	// Initialize HttpClient
-	httpClient = &http.Client{}
-
-	// Initialize Firestore
-	client, err := firestore.NewClient(context.Background(), currentProject)
-	if err != nil {
-		return fmt.Errorf("CreateUser [Init Firestore]: %v", err)
-	}
-
-	firestoreClient = client
-
-	return nil
-}
-
-// generateAppleMusicReq created the request object to call Apple Music API
-func generateAppleMusicReq(uri string, method string, devToken string) (*http.Request, error) {
-	// Generate request
-	appleMusicReq, appleMusicReqErr := http.NewRequest(method, uri, nil)
-	if appleMusicReqErr != nil {
-		return nil, fmt.Errorf("GetAppleMusicMeda [http.NewRequest]: %v", appleMusicReqErr)
-	}
-
-	// Add headers
-	appleMusicReq.Header.Add("Authorization", "Bearer "+devToken)
-
-	return appleMusicReq, nil
-}
-
-// getAppleMusicTrack will call Apple Music Song API to get the metadata for a song
-func getAppleMusicTrack(trackID string, storefront string, appleDevToken firebase.FirestoreAppleDevJWT) (*firebase.FirestoreMedia, error) {
-	// Generate request
-	appleMusicGetTrackReq, appleMusicGetTrackReqErr := generateAppleMusicReq(catalogHostname+"/"+storefront+"/songs/"+trackID, "GET", appleDevToken.Token)
-	if appleMusicGetTrackReqErr != nil {
-		return nil, fmt.Errorf("GetAppleMusicMeda [http.NewRequest]: %v", appleMusicGetTrackReqErr)
-	}
-
-	// Make request
-	getTrackResp, getTrackRespErr := httpClient.Do(appleMusicGetTrackReq)
-	if getTrackRespErr != nil {
-		return nil, fmt.Errorf("GetAppleMusicMedia [client.Do]: %v", getTrackRespErr)
-	}
-
-	// Check to see if request was valid
-	if getTrackResp.StatusCode != http.StatusOK {
-		// Convert Apple Music Error Object
-		var appleMusicErrorObj social.AppleMusicRequestError
-
-		err := json.NewDecoder(getTrackResp.Body).Decode(&appleMusicErrorObj)
-		if err != nil {
-			return nil, fmt.Errorf("GetAppleMusicMedia [Apple Music Request Decoder]: %v", err)
-		}
-
-		return nil, fmt.Errorf("GetAppleMusicMedia [Apple Music Track Request]: %v", appleMusicErrorObj)
-	}
-
-	var appleMusicTrackData appleMusicTrackResp
-	respDecodeErr := json.NewDecoder(getTrackResp.Body).Decode(&appleMusicTrackData)
-	if respDecodeErr != nil {
-		return nil, fmt.Errorf("GetAppleMusicMedia [Apple Music Response Decoder]: %v", respDecodeErr)
-	}
-
-	log.Println(appleMusicTrackData)
-
-	// Check for length to make sure we found a match
-	if len(appleMusicTrackData.Data) == 0 {
-		return nil, fmt.Errorf("GetAppleMusicMedia [Length Check]: No results in data for id %s", trackID)
-	}
-
-	// Setup FirestoreMeida object
-	track := appleMusicTrackData.Data[0]
-	firestoreMedia := firebase.FirestoreMedia{
-		Name:    track.Attributes.TrackName,
-		Album:   track.Attributes.AlbumName,
-		Type:    "track",
-		Creator: track.Attributes.ArtistName,
-		Apple: firebase.FirestoreMediaPlatformData{
-			ID:     trackID,
-			URL:    track.Attributes.ExternalURL,
-			Images: track.Attributes.Artwork,
-		},
-	}
-
-	return &firestoreMedia, nil
-}
-
-// getSpotifyPlaylist will call Apple Music Playlist API to get the metadata for a playlist
-func getAppleMusicPlaylist(playlistID string, storefront string, appleDevToken firebase.FirestoreAppleDevJWT) (*firebase.FirestoreMedia, error) {
-	// Generate request
-	appleMusicGetPlaylistReq, appleMusicGetTrackReqErr := generateAppleMusicReq(catalogHostname+"/"+storefront+"/playlists/"+playlistID, "GET", appleDevToken.Token)
-	if appleMusicGetTrackReqErr != nil {
-		return nil, fmt.Errorf("GetAppleMusicMeda [http.NewRequest]: %v", appleMusicGetTrackReqErr)
-	}
-
-	getPlaylistResp, getPlaylistRespErr := httpClient.Do(appleMusicGetPlaylistReq)
-	if getPlaylistRespErr != nil {
-		return nil, fmt.Errorf("GetAppleMusicMedia [client.Do]: %v", getPlaylistRespErr)
-	}
-
-	// Check to see if request was valid
-	if getPlaylistResp.StatusCode != http.StatusOK {
-		// Convert Apple Music Error Object
-		var appleMusicErrorObj social.AppleMusicRequestError
-
-		err := json.NewDecoder(getPlaylistResp.Body).Decode(&appleMusicErrorObj)
-		if err != nil {
-			return nil, fmt.Errorf("GetAppleMusicMedia [Apple Music Request Decoder]: %v", err)
-		}
-
-		return nil, fmt.Errorf("GetAppleMusicMedia [Apple Music Track Request]: %v", appleMusicErrorObj)
-	}
-
-	var appleMusicPlaylistResp appleMusicPlaylistResp
-	respDecodeErr := json.NewDecoder(getPlaylistResp.Body).Decode(&appleMusicPlaylistResp)
-	if respDecodeErr != nil {
-		return nil, fmt.Errorf("GetAppleMusicMedia [Apple Music Response Decoder]: %v", respDecodeErr)
-	}
-
-	log.Println(appleMusicPlaylistResp)
-
-	// Check for length to make sure we found a match
-	if len(appleMusicPlaylistResp.Data) == 0 {
-		return nil, fmt.Errorf("GetAppleMusicMedia [Length Check]: No results in data for id %s", playlistID)
-	}
-
-	// Setup FirestoreMedia object
-	playlist := appleMusicPlaylistResp.Data[0]
-	firestoreMedia := firebase.FirestoreMedia{
-		Name:    playlist.Attrbutes.Name,
-		Album:   "Playlist",
-		Type:    "playlist",
-		Creator: playlist.Attrbutes.CuratorName,
-		Apple: firebase.FirestoreMediaPlatformData{
-			ID:     playlistID,
-			URL:    playlist.Attrbutes.URL,
-			Images: playlist.Attrbutes.Artwork,
-		},
-	}
-
-	return &firestoreMedia, nil
+	writer.WriteHeader(http.StatusOK)
+	writer.Header().Set("Content-Type", "application/json")
+	json.NewEncoder(writer).Encode(firestoreMediaData)
 }
